+++ Packages/com.vareiko.foundation/Tests/Runtime/Ads/SimulatedAdsServiceTests.cs
@@ -127,6 +127,7 @@ namespace Vareiko.Foundation.Tests.Ads
             container.DeclareSignal<AdShownSignal>();
             container.DeclareSignal<AdShowFailedSignal>();
             container.DeclareSignal<AdRewardGrantedSignal>();
+            container.DeclareSignal<AdsOperationTelemetrySignal>();
             return container.Resolve<SignalBus>();
         }